# rpi5_snapshot/openapi.yaml
openapi: 3.0.3
info:
  title: Game Stats API
  version: 1.0.0
  description: API kontrakty dla modułów 1v1 oraz multiplayer.
servers:
  - url: /api/v1
components:
  schemas:
    Uuid:
      type: string
      format: uuid
      example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
    Date:
      type: string
      format: date
      example: "2026-01-30"
    DateTime:
      type: string
      format: date-time
      example: "2026-01-30T12:00:00.000Z"
    ErrorDetailsItem:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
    ErrorPayload:
      type: object
      required: [code, message, details]
      properties:
        code:
          type: string
          enum: [VALIDATION_ERROR, NOT_FOUND, CONFLICT, INTERNAL_ERROR]
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetailsItem'
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorPayload'
...
    Game:
      type: object
      required: [id, code, name, isActive]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        code:
          type: string
        name:
          type: string
        isActive:
          type: boolean
...
    MatchSummary:
      type: object
      required: [id, playedOn, game, playerA, playerB, scoreA, scoreB, winner]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        playedOn:
          $ref: '#/components/schemas/Date'
        game:
          $ref: '#/components/schemas/Game'
        playerA:
          type: object
          required: [id, name]
          properties:
            id:
              $ref: '#/components/schemas/Uuid'
            name:
              type: string
        playerB:
          type: object
          required: [id, name]
          properties:
            id:
              $ref: '#/components/schemas/Uuid'
            name:
              type: string
        scoreA:
          type: integer
        scoreB:
          type: integer
        winner:
          type: string
          enum: [A, B, DRAW]
    MatchDetail:
      allOf:
        - $ref: '#/components/schemas/MatchSummary'
        - type: object
          required: [notes, createdAt, updatedAt]
          properties:
            notes:
              type: string
              nullable: true
            createdAt:
              $ref: '#/components/schemas/DateTime'
            updatedAt:
              $ref: '#/components/schemas/DateTime'
    MatchListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MatchSummary'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    MatchCreate:
      type: object
      required: [gameId, playedOn, playerAId, playerBId, scoreA, scoreB]
      properties:
        gameId:
          $ref: '#/components/schemas/Uuid'
        playedOn:
          $ref: '#/components/schemas/Date'
        playerAId:
          $ref: '#/components/schemas/Uuid'
        playerBId:
          $ref: '#/components/schemas/Uuid'
        scoreA:
          type: integer
          minimum: 0
        scoreB:
          type: integer
          minimum: 0
        notes:
          type: string
          nullable: true
          maxLength: 2000
    MatchPatch:
      type: object
      properties:
        gameId:
          $ref: '#/components/schemas/Uuid'
        playedOn:
          $ref: '#/components/schemas/Date'
        playerAId:
          $ref: '#/components/schemas/Uuid'
        playerBId:
          $ref: '#/components/schemas/Uuid'
        scoreA:
          type: integer
          minimum: 0
        scoreB:
          type: integer
          minimum: 0
        notes:
          type: string
          nullable: true
          maxLength: 2000
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Internal error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
...
  /matches:
    get:
      summary: List 1v1 matches
      parameters:
        - in: query
          name: gameId
          schema:
            $ref: '#/components/schemas/Uuid'
        - in: query
          name: playerId
          schema:
            $ref: '#/components/schemas/Uuid'
        - in: query
          name: dateFrom
          schema:
            $ref: '#/components/schemas/Date'
        - in: query
          name: dateTo
          schema:
            $ref: '#/components/schemas/Date'
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
            minimum: 0
        - in: query
          name: sort
          schema:
            type: string
            enum: [playedOnDesc]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Create 1v1 match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /matches/export.csv:
    get:
      summary: Export 1v1 matches to CSV
      parameters:
        - in: query
          name: gameId
          schema:
            $ref: '#/components/schemas/Uuid'
        - in: query
          name: playerId
          schema:
            $ref: '#/components/schemas/Uuid'
        - in: query
          name: dateFrom
          schema:
            $ref: '#/components/schemas/Date'
        - in: query
          name: dateTo
          schema:
            $ref: '#/components/schemas/Date'
      responses:
        '200':
          description: CSV export
          content:
            text/csv:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /matches/{id}:
    get:
      summary: Get 1v1 match by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: '#/components/schemas/Uuid'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      summary: Update 1v1 match
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: '#/components/schemas/Uuid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchPatch'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      summary: Delete 1v1 match
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: '#/components/schemas/Uuid'
      responses:
        '204':
          description: No content
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
...